combine shared operations tests
